requirements.txt
# requirements.txt
streamlit>=1.29.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-groq>=0.0.3
unstructured[all-docs]>=0.11.0
pillow>=10.0.0
lxml>=4.9.0
chromadb>=0.4.0
tiktoken>=0.5.0